move logger interface to outer log package
