+++ internal/cloud/cloud_discovery.go
@@ -24,8 +24,8 @@ import (
 	"strings"
 
 	"github.com/hazelcast/hazelcast-go-client/cluster"
-	"github.com/hazelcast/hazelcast-go-client/internal/logger"
 	"github.com/hazelcast/hazelcast-go-client/internal/rest"
+	"github.com/hazelcast/hazelcast-go-client/logger"
 )
 
 const envCoordinatorBaseURL = "HZ_CLOUD_COORDINATOR_BASE_URL"
